Get it building
To get a shape of the thing, I started by getting to a point where I could build the main application, whose entry point (WinMain) can be found in blades.cpp.
Then, in a loop I:
* Fix any compiler errors with minimal changes I can get away with.
* Attempt to link and notice the names of any missing symbols. Search for those symbols in the source and add in the source file that seems to contain it.
* Go back to the compile phase.

Eventually I reach the end of the cycle where the only missing symbols are Windows API symbols, at which point we add any library dependencies to take care of those (just Winmm.lib in this case).

Attempting to run the application, not unsurprisingly leads to a hanging application, but the object here was just to get us setup with a successful compile.

A number of issues presented themselves in the source:
1) Non const character arrays being initialized with const data and non const character arrays being passed to functions which desire const character arrays.
2) Keywords which no longer have any meaning and upset the compiler: pascal, PASCAL, _export. Some, like "huge" and "FAR" don't have a meaning since we got a flat memory model (i.e. post 16-bit) but they tend to be more harmless.
3) min and max functions. This has been a thorn in the side of many engineers attempting to write cross platform code over the years as it wasn't always available or was under a different name, or had conflicting macros. We need to remove the definitions used here since they will now be well replaced with the commonly available macros.
4) Variables/constants that share the same name as things in the Windows header files: rad2 and rad3.
5) Windows 16-bit API functions/macros which are no longer available. MAKEPOINT, SetViewportOrg, GetFreeSystemResources, MoveTo
6) More precise function prototypes. As the Win32 GUI functionality involved, the headers got tightened up to try and provide some amount of type safety, although many would feel the horse had already bolted here. Specifically the Dialog and Window functions that are registered when these are created need to match a stronger signature of DLGPROC or WNDPROC now.
7) Win32 GDI objects were more or less just HANDLE objects back in the day, and C wouldn't really mind if you swapped one for another. C++ is rather more intolerant of this, so we either need to cast types or find other ways to get to work around these incompatibilities.
8) In, I think one place only (WinMain), there was actually a K&R function definition syntax.
9) Some Win32 GUI APIs can take (or perhaps used to take?) integer arguments instead of objects. C didn't mind, but C++ does, so we need to cast those int's to the expected types. These were typically HMENU handles.
10) More Win 16-bit GUI issues come from calls to GetWindowWord which are used for various purposes, all strongly discouraged for Win32 GUI programming, but they are still present even though some of the defines have either changed names or been removed. These were typically identifiers beginning GWW_ and occasionally for GetWindowLong a GWL_ identifier.
11) More 16-bit GUI things removed from 32-bit was the use of the OF_SEARCH flag being passed to "OpenFile", so we'll have to see what that was being used for and if we need to work around it.
12) The FSClose function does not return a value even though the signature shows a short.
13)  An interesting oddity in "kill_pc" which the compiler tells us is a bit odd: we're comparing "adven[i].status > 0" but the compiler says that status is an array of shorts rather than a value. A quick look around might indicate that "main_status" should have been used instead, but it's too early to say so this should probably have an assert added to make sure we debug the first time it crops up.
14) One of the most unusual patterns to spot is where we see an incompatible type being substituted for another type: for example an HWND or HDC being passed as a HBITMAP. These can be difficult to spot without some background in Win32 GDI programming, because sometimes we do indeed have to do strange casts, but they are more common with types that take a HANDLE or HGDIOBJ. Since in our cases these cannot be treated interchangeably, we need to dig a little deeper and we discover that these functions *also* take a parameter which describe the type of object and then get handled differently in that function. A sort of poor man's polymorphism. I guess the idea was to share some degree of common code/pattern, but it does make our job harder as we're trying to detect bad conversions. For now these will be cast to the expected type as per the original pattern, but C++ function overloading should come to our rescue when we look at improving the situation incrementally without losing the type at the call site.

A brief attempt to debug the initial app startup shows that the first things that are failing are that we couldn't load any app resources, and this isn't surprising as we haven't added any. BLADES.RC seems to be the likely candidate so adding that to the project got us a little further and we get a window with a cursor. There was a missing file referenced in BLADES.RC called "look.cur" which I have temporarily replaced with "boot.cur" (which I think is otherwise unreferenced) but we'll have to see if that's even loaded later.
So we get a bit further this time, but still the hang very quickly afterwards. Breaking into the app at this point with a debugger revealed we are stuck waiting on a SendMessage call to complete which never does. This may well be some GUI behaviour that has changed over the years, but the reason we're hanging is that we have a single threaded application that has created a WndProc to handle Windows messages but then goes on to call SendMessage before it gets to the point where it sets up the standard GetMessage/DispatchMessage loop. The problem with this is that SendMessage will block waiting for the message to be handled, and it will never be handled. Changing it to a PostMessage gets us a fair bit further: the cursor appears, music plays and the menu's do... something, albeit we can't see what - but we occasionally hear a noise when we click on them :)
This is probably as far as it's sensible to get right now. We've probably got what we need to deduce our first set of refactorings to get us to a point where we can build the app.
Note that we haven't looked at any warnings or informationals yet and we haven't looked at setting higher warning levels or anything else. Just a base point of compiling it.