Initial CMake build project along with simple fixes which prevent compilation (#1)
The objective for this checkin is to have all files compiling and getting to the end of the compilation without stopping. i.e. if a syntax error prevented further compilation then I tried to resolve it. This isn't the same as compiling without errors.

The errors I made changes to avoid should all be zero consequence - i.e. they are not logic changes:
* Remove the lowercase "pascal" as it is defined to be __stdcall on Windows and this prevented lines such as the following from being understandable to the compiler: pascal Boolean myfunction();
* Remove the "_export" as it is not defined.
* Remove "huge" as it is not defined. It probably related to memory models from those 16-bit days.
* Rename the global variables "rad2" and "rad3" as they conflict with some #define's in Windows headers.
* Move "WndProc" and "WinMain" away from K&R style function syntax.
* Remove #include of missing headers "small.h", "medium.h" and "large.h".

The following error has some consequence, given that a different cursor will be used:
* Adjust paths in BLADES.RC to correctly find files. Also deal with the missing "look.cur" by leveraging the unreferenced "boot.cur" file which is present. Will need to resolve this in future. Also include Windows header for "VK_F1" definition.