draw_dialog_graphic variants draw_dialog_graphic_bmp and draw_dialog_graphic_wnd
The function "draw_dialog_graphic" currently takes a single type of HWND, but the last parameter actually tells the function to treat it as HWND (0) or a HBITMAP (1).
Although we could have provided identically named functions that deduce the type (C++ overloads) we'd have to have been very careful because these variables are often being cast at the call site, so at this stage I instead introduced some simple forwarding functions with types and names that match, and removed the parameter.
So that means a call like this:
    draw_dialog_graphic((HWND) talk_gworld, face_rect, 1000 + i, FALSE,1);
Becomes:
    draw_dialog_graphic_bmp(talk_gworld, face_rect, 1000 + i, FALSE);
I don't expect it all to stay this way forever, but for now it's a relatively simple diff.