Move from rect_draw_some_item to rect_draw_some_item_* variants
Add variants to support: HWND, HDC, BMP and the either function where a flag indicates which. Again, very much a work in progress - the objective being that we keep the casts in the fewest places possible so that when we try and unpick them later we can hopefully find a clear refactoring route.
I also moved the helper functions I've been adding into their own headers. These are stopgap measures, I expect them to disappear in a later refactor - but since this functionality spans the app, the scenario editor and the party editor - I wanted them in one place.