Refactor variant gui api calls (#4)
* draw_dialog_graphic variants draw_dialog_graphic_bmp and draw_dialog_graphic_wnd

The function "draw_dialog_graphic" currently takes a single type of HWND, but the last parameter actually tells the function to treat it as HWND (0) or a HBITMAP (1).
Although we could have provided identically named functions that deduce the type (C++ overloads) we'd have to have been very careful because these variables are often being cast at the call site, so at this stage I instead introduced some simple forwarding functions with types and names that match, and removed the parameter.
So that means a call like this:
    draw_dialog_graphic((HWND) talk_gworld, face_rect, 1000 + i, FALSE,1);
Becomes:
    draw_dialog_graphic_bmp(talk_gworld, face_rect, 1000 + i, FALSE);
I don't expect it all to stay this way forever, but for now it's a relatively simple diff.

* Remove duplicate header includes

This will later be fixed by sorting out these headers, adding pragma once, or similar, but for now let's just resolve the issues by not including the same header twice.

* Move from rect_draw_some_item to rect_draw_some_item_* variants

Add variants to support: HWND, HDC, BMP and the either function where a flag indicates which. Again, very much a work in progress - the objective being that we keep the casts in the fewest places possible so that when we try and unpick them later we can hopefully find a clear refactoring route.
I also moved the helper functions I've been adding into their own headers. These are stopgap measures, I expect them to disappear in a later refactor - but since this functionality spans the app, the scenario editor and the party editor - I wanted them in one place.

* Move from paint_pattern to variants

Variants are: HBITMAP, DC, WND and one where no dest is passed, which should cause it to use a global mainPtr.