SelectObject return type refactoring
The GDI32 "SelectObject" function returns a GUI handle to the previously owned object, in case it needs to be released or used again later. These don't appear to be reused, but they are cached, so for now let's let the cached object be of the type "SelectObject" returns: HGDIOBJ to avoid unnecessary casting.