Move away from retired constants
GWW_ID, GWL_WNDPROC and GWW_HINSTANCE are all retired largely because as the width of things increased first to 32-bit and then 64-bit, these clearly could not fit pointers inside them any more. The GWW_ID one might still have been okay as I believe it may return a small identifier rather than a pointer, but still.
This is a speculative attempt to write a replacement, but it is a risk which I will have to create a ticket to look back into later.