Tidy up missing identifiers and remaining 16-bit only GDI functions/macros replaced (#9)
* OF_SEARCH no longer exists (or is supported)

We would need more work to try and work out how to simulate the missing behaviour of the OS_SEARCH parameter, so we'll comment it out and create an issue in investigate it later.

* Move away from retired constants

GWW_ID, GWL_WNDPROC and GWW_HINSTANCE are all retired largely because as the width of things increased first to 32-bit and then 64-bit, these clearly could not fit pointers inside them any more. The GWW_ID one might still have been okay as I believe it may return a small identifier rather than a pointer, but still.
This is a speculative attempt to write a replacement, but it is a risk which I will have to create a ticket to look back into later.

* GetFreeSystemResources API is removed

Created a ticket to track that these are gone in case I can find something useful to put in, but these are only displayed for diagnostic purposes.

* GDI API MoveTo no longer exists

Interesting background as to why they didn't just provide a macro to call MoveToEx directly here: https://devblogs.microsoft.com/oldnewthing/20090720-00/?p=17433

* Replace MAKEPOINT macro calls

MAKEPOINT was removed from the GDI headers so we need to pull it apart ourselves (or I could have used MAKEPOINTS instead, but it would involve larger changes than I'd like to make at this time.

* Missed the changes to the GW* constant changes for editor source

* Remaining missing 16-bit GDI APIs replaced

SetViewportOrg ==> SetViewportOrgEx
GetTextExtent borrowed some code from Wine, which should get us where we need to get for now.