The first questionable decision is calling GetDC
GetDC has always confused me, along with most of the rest of the Windows GDI. I've taken a chance here and made a couple of calls to GetDC even though one of them really doesn't make much sense to me, but I assume the original was working. This is definitely something I'm going to have to come back to.