The second questionable change: No SendMessage
If you run the app without this change then it will hang on the call to SendMessage.
Debugging through the code and I can see why: it's calling SendMessage on the main thread before the DispatchMessage loop is underway, so I think this means it will hang as there is no way for this single threaded app to receive and respond to its own message.
I've changed this to a PostMessage call in the places might do this during initial setup. I will raise an issue to reconsider this after I have a better understanding.