Two questionable decisions (#13)
* The first questionable decision is calling GetDC

GetDC has always confused me, along with most of the rest of the Windows GDI. I've taken a chance here and made a couple of calls to GetDC even though one of them really doesn't make much sense to me, but I assume the original was working. This is definitely something I'm going to have to come back to.

* The second questionable change: No SendMessage

If you run the app without this change then it will hang on the call to SendMessage.
Debugging through the code and I can see why: it's calling SendMessage on the main thread before the DispatchMessage loop is underway, so I think this means it will hang as there is no way for this single threaded app to receive and respond to its own message.
I've changed this to a PostMessage call in the places might do this during initial setup. I will raise an issue to reconsider this after I have a better understanding.