townout.cpp is missing
Well, not quite true, but it was missing and Jeff furnished us with the macOS version alongside some helpful comments about ways to progress at the top of townout.cpp.
For now, I've disabled the contents and replaced the one reference to a macOS type "Str255" (which is just a typedef to a 256 byte character array) with a straight character pointer.