Fix resource files
Both needed the include of WinUser.h for the definition of the virtual key macro.
Also change the paths to the sound files to find the right things.
The scene editor must have been created in a time when the bytes could be directly provided for things like cursors. However they haven't been for some time, so I saved the bytes into .cur files and referenced them from the .rc file.