Build editors (#19)
* Add CMake files to build the editors

These are actually straight from the Experiment-2 branch since I know they will work once changes to the source are made.

* townout.cpp is missing

Well, not quite true, but it was missing and Jeff furnished us with the macOS version alongside some helpful comments about ways to progress at the top of townout.cpp.
For now, I've disabled the contents and replaced the one reference to a macOS type "Str255" (which is just a typedef to a 256 byte character array) with a straight character pointer.

* Replace macOS Rect type for Windows RECT type

The declaration of flip_rect in the header takes a "Rect" whilst the defintion takes a "RECT", so change the header to match the Windows type.

* make_new_scenario declaration did not match definition

So update the declaration to match the definition (and the one call site).

* Explicit casts

1) SelectStockObject with a brush type explicitly cast to be a HBRUSH where they are assigned to some Windows structure.
2) GlobalLock call returns cast to the expected type.

* Port 16-bit GDI API calls to 32-bit equivalents

Ported:
* SetViewportOrg ==> SetViewportOrgEx
* MoveTo ==> MoveToEx
* GetTextExtent ==> a clone of the Wine function which forwards the call.

* FS* API issues

A few function declarations were missing from a header in one case, and in the other (like elsewhere) the FSClose function was not returning a value.

* Include helper header

These were from some imperfect changes I made trying to keep the different versions in sync but before I could actually compile the editors.

* SelectObject store variable types to HGDIOBJ

Since these objects are only used in our context potentially for later calls to SelectObject, change the type to match the return from SelectObject.

* Missing assert header and missing function declaration

* Missed a GWL_WNDPROC ==> GWLP_WINDPROC

In my pre-compile changes, I missed this one!

* Same change as elsewhere for now WRT resource ids to HMENU handles

Hide the pain for now, to evaluate later.

* GetMapMode as per previous changes

Probably a mistake, but we'll need to investigate this later.

* cd_set_focus takes a parameter in the scenario editor

All other examples of this take no parameter at all, except for the scenario editor - so this will need digging into. Have raised an issue.

* handle_keystroke should return a value

But it does not... For now add an assert and return FALSE.

* Fix resource files

Both needed the include of WinUser.h for the definition of the virtual key macro.
Also change the paths to the sound files to find the right things.
The scene editor must have been created in a time when the bytes could be directly provided for things like cursors. However they haven't been for some time, so I saved the bytes into .cur files and referenced them from the .rc file.

* Provide stopgap townout source

Until I have the time to work out what this should look like, I'll provide a version that just stubs out the functions and use that instead.

* Change the type of global "give_delays" declation to match definition