Add warning suppressions for Clang-cl
clang-cl is a strange creature "neither fish nor fowl", as they say.
However it does do a fantastic job at detecting code smells and hopefully prepare us for the eventual point where we can build parts of the code for a platform other than Windows.
I should probably avoid setting both Windows warning flags and clang-cl ones, but for now we're still at the: get it to build.
There are two lists of warnings, the first looked like they might take some effort to address and perhaps aren't as serious as the second set.
The second set are either serious enough to do something about or trivial enough to address easily.
One bug came out which is described here: https://github.com/CorrodedCoder/Blades-of-Exile/issues/20.
The workaround for now is to assert so that we can evaluate it at runtime and leave a comment about the mystery.