System header anglebracket mandate
Although one is free to choose, typically system headers are included in angle brackets. This will become more important when I switch on warning level 4 for Windows as I can tell it to ignore anglebracket included files in that regard, which makes it functional (i.e. we won't see warnings from Windows headers).
For now, make sure they are angle bracketed and whilst I'm at it, let's state the intention cleanly that we're in a C++ world by using, for example <cstdio> instead of <stdio.h>