Cleanups (#22)
* Rename Global.h to global.h

There was a large mismatch on which files included "Global.h" and which "global.h" but in the end the lowercase won it, just because all of the rest of the headers are lowercase.

* Deal with sprintf/sscanf format issues

The ones involving sprintf were harmless but the sscanf ones were definitely an issue since they were trying to deserialize a string into a short but passing %d which treated the variable as an int. When I ran in the debugger I saw stack corruption in one case.

* Remove line continuation character from comments

* System header anglebracket mandate

Although one is free to choose, typically system headers are included in angle brackets. This will become more important when I switch on warning level 4 for Windows as I can tell it to ignore anglebracket included files in that regard, which makes it functional (i.e. we won't see warnings from Windows headers).
For now, make sure they are angle bracketed and whilst I'm at it, let's state the intention cleanly that we're in a C++ world by using, for example <cstdio> instead of <stdio.h>

* Remove min/max functions

They come from C/C++ now.

* Remove MSVC unreferenced local variable warning

So this took a lot of effort, there were many hundreds of these - there is probably some tooling out there to make this easier!