Rework warning levels
We ideally want a clean build with max warnings but of course to do that with old code we need to suppress a hell of a lot of warnings.
The idea is that we'll uncomment the suppressions one at a time and take care of the relevant issues until we have a more meaningful list.
We also provide  max warnings mode with almost no warnings suppressed, and this can be enabled by setting BOE_WARNINGS_MAX.