Cleanup 2 (#26)
* Remove GlobalAlloc calls

I'm guessing this was a compiler, linker or executable problem back in the day, but no longer necessary.

* Provide wrapper around LoadCursor

The idea here is to start capturing failures to load resources so we don't silently ignore them.

* Replace struct and array with std::array of std::array

Less code and at some point we'll see about getting some better filling of arrays without risk of overflow.

* Move Text external functions to static where possible

Trying to avoid exposing functions which aren't used by other source files. In some cases it's as simple as removing the function declaration from the header and making it static.
In others to achieve the above we sometimes need to move a function to a different source file.

* Get the game build functional

Add missing RC files to game build, also one was missing the inclusion of WinUser.h for virtual key definition.
Copy the support files needed for the application to run to the output directory (scenarios and BMP files).