Change type of RECT in persisted game structures to BoeRect
By changing the type, we resolve https://github.com/CorrodedCoder/Blades-of-Exile/issues/28 and allow scenarios to be read from a file. With this change we can now load and play a scenario, at least to some  degree!

There are, at least temporarily, now two versions of alter_rect. The old one is still used by Windows code, I think to handle RECT structures initialized in Mac rect format. This could be resolved by making the changes to the source so that the data is no longer in Mac rect format.