Extract common endian adjustment to single place
Extract endian byte swapping into one single source file so that we can leverage it from multiple places. We leave the decision making (i.e. whether to byte swap or not) in the calling code as a separation of concerns.
Both the scenario editor and the game now share the same source.