Add DumpScenario tool
This is useful to make sense of scenario data. It is not yet complete.
This should be cross platform so I needed to remove reference to the "far" modifier (which isn't meaningful anyway).
I also tidied up the includes for endian adjust for C++23 as the includes are only needed for C++20 compiles.