Load scenarios (#29)
* Windows scenario files are corrupted

The Windows scenario files are corrupted as detailed here https://github.com/CorrodedCoder/Blades-of-Exile/issues/27

The good news is that the Mac ones seem okay, so let's switch to using those instead.

* Remove shortloc from various headers into implementation

The shortloc structure is only used in one source file, so move this inside that source file and remove APIs that reference it from headers.

* Consolidate identical global.h into a single file

Character Editor had an identical copy of global.h so let's make them share the same file so we don't need to keep making the same changes twice.

* Re-order structs ahead of extraction to common header

This struct is different between scenario editor and the game itself, I'm guessing because it's only in memory for the scenario editor, as seems to be similar for other structs.

* Extract common structure definitions into common header file

These scenario, town, etc structures are common across all uses and specifically are read into memory from disk and written back from memory, so the layout must not change.

* Make flip_short internal

This function is only used inside one translation unit (well, in each project - we'll fix that later) so let's make it visible only inside that file.

* Make flip_rect internal and alter_rect where possible

flip_rect, like flip_short before it, is well contained in a single source file with no external references, so make that explicit.
Do the same for alter_rect where possible, but there is one special case coming up where we will eventually end up with two different versions of this.

* Change type of RECT in persisted game structures to BoeRect

By changing the type, we resolve https://github.com/CorrodedCoder/Blades-of-Exile/issues/28 and allow scenarios to be read from a file. With this change we can now load and play a scenario, at least to some  degree!

There are, at least temporarily, now two versions of alter_rect. The old one is still used by Windows code, I think to handle RECT structures initialized in Mac rect format. This could be resolved by making the changes to the source so that the data is no longer in Mac rect format.

* Remove last traces of alter_rect for RECT types on Windows

We provide a compatibility macro for initializing RECT structures with constant data so that it can stay in BOE's preferred format and be handled correctly in both places. This is probably short term, but at least we can get away without the runtime alter_rect.

* Add static asserts to ensure data structure sizes are correct

This way we should find out early if there's some kind of mismatch from what we're expecting.

* Extract common endian adjustment to single place

Extract endian byte swapping into one single source file so that we can leverage it from multiple places. We leave the decision making (i.e. whether to byte swap or not) in the calling code as a separation of concerns.
Both the scenario editor and the game now share the same source.

* Add DumpScenario tool

This is useful to make sense of scenario data. It is not yet complete.
This should be cross platform so I needed to remove reference to the "far" modifier (which isn't meaningful anyway).
I also tidied up the includes for endian adjust for C++23 as the includes are only needed for C++20 compiles.

* Clean up endian_adjust by flattening out loops via helper

Break out each distinct type into it's own function and leverage a fixed array helper to do the right thing without ugly for loops.