Cleanup 3 (#30)
* Provide convenience wrapper around endian adjust

I think it looks a little nicer, although it *might* have increased the code size a little.

* Add convenient controls to CMake script

Support easily enabling debug symbols for release builds and also to enable the address sanitizer (recommended!).

* General tidy up

1. Remove unreferenced global/static variables.
2. Changed some arrays to std::array's.

* Remove "far" modifier

Legacy thing, not needed.

* Remove confusing an unnecessary global variable "i"

Also make a few other variables static.

* Another unreferenced variable removed (thanks clang!)

* Remove missing-prototypes warning suppression from clang

This is helpful because it correctly pointed out functions that could be declared static, and in some cases this allowed functions to be removed entirely since they were clearly no longer referenced.
Now truthfully making functions static is a bit of a fudge because this is more of a C thing to do that C++. In C++ we should be placing functions not required outside of this translation unit (source file) into the anonymous namespace. I don't want to do that for the existing code yet because it'll be a big change that will be hard to track (lots of reorganisation in the same files)... however...
It should have been possible to do this for my own new C++ files, but I confess that my template knowledge is letting me down here: when I move the functions into the null namespace clang (on Linux), although not on Windows, and VC++ get quite cross with me.
Speaking of which, clang-cl 16 doesn't know about the wmain function so it complains of a missing prototype. I've suppressed that directly in the source.
I also made the DumpScenario project buildable on Linux with clang 17 and libc++, although it took some jumping through hoops to work out how via WSL. It seems that clang 17 or possibly libc++ is better at keeping honest with respect to including the right system headers so I added in a few which were missing.

* Remove unnecessary references to mainPtr

* Remove unnecessary extern reference to startup_loaded or the variable itself in some places

* Remove one last "far" no longer necessary modifier

* Make some non static functions static

Mostly these are implementation detail functions only used in a single source file, so no reason for them to be part of the API.
Some functions are no longer referenced at all: swap_val, change_val_4, change_val.

* Make all file handling functions static as they are internal to the various forms of fileio source

* Move port* functionality inline ahead of refactor

Should be a no-operation change.