Combine action and error check
Somewhat ironic that I tried to untangle these earlier, but in essence I've realised it will be easier to see the diff when I finally move to to fstreams if the comparison is a one liner.
E.g. I don't need to know the resturn type of file_read_type vs llfile_read_type because they are all in the if statement, so I will just replace them with a function bool condition.
I probably should have gone deeper in my experimental refactoring of the filesystem before charging ahead...