Finally switch across to a common pattern for error handling
The diff requires a big font, but hopefully it's easier to see the change than it would have been if this was all combined into one.
Of course there are many paths here which seem to ignore errors.
Sadly we'd need to do a fair bit more work before an exception might be thrown in case of bad file reading, so we must go slowly...