Wrap xor encoding of element to auto decode
When the save game is saved to disk, some parts of it are obfuscated using a simple xor. These are xor'd before writing but because they are live data they must be xor'd again to get back the original values afterwards, including on the error path.
This means specifying the xor three separate times and of course you have to be careful to do so.
I created a simple object to manage this which is created (does the first xor), has a  method called get which returns the underlying xor'd structure we want to serialize, and repeats the xor on destruction (i.e. after the serialize call has finished), leaving the data unobfuscated again.