Simplify the read by enabling exceptions
Introduced a stream reader which requires that the input stream throws on read failures.
Small behaviour change: attempting to read a game save file shorter than 6 bytes used to beep, but no error dialog popped up. The new behaviour is the same as all other types of read failure in that it beeps and then pops up a dialog box saying it couldn't load the save file.