Finally address the dialog box size issues
As I mention in the code comment, this is no great understanding on my part: just trial and error and comparing against Windows 3.1 screenshots. I should try and get to the bottom of this but I don't understand the dialog framework nearly well enough to do that yet.
Tracked under https://github.com/CorrodedCoder/Blades-of-Exile/issues/32