paint_pattern gains some type safety
The simplest one to change first, the pattern is that by passing a variant and asking for the type we believe it to be at the right time, we can validate that it's the type we were expecting. i.e. if a HBITMAP is passed to a function expecting a HDC then we would have to call std::get<HDC> which would throw an exception.
Eventually we can propagate these variants onwards so that we no longer need to check until the point of use, but since there are so many other casts already in play, we need to do this slowly and deliberately so that, to the best of my knowledge, all callers of paint_pattern are now passing their true type, rather than one which it has already been cast to.