Unsafe graphics types refactored (#36)
* Move const source rects out of function

* Split draw_pc_effects into two halves ahead of refactor

* Adjust signature to new function to take specific adventurer

* Re-order helper function and make it static

* Expand ahead of moving to two functions

* Split into two helper functions

* Tidy up helper functions

* Get rid of the -10 offset business and limit the type safety to a single dynamic cast

* Remove unnecessary paint_pattern_wnd as it's already covered by paint_pattern_main

When 1 is passed to the underlying paint_pattern it ignores this input parameter in any case.

* paint_pattern gains some type safety

The simplest one to change first, the pattern is that by passing a variant and asking for the type we believe it to be at the right time, we can validate that it's the type we were expecting. i.e. if a HBITMAP is passed to a function expecting a HDC then we would have to call std::get<HDC> which would throw an exception.
Eventually we can propagate these variants onwards so that we no longer need to check until the point of use, but since there are so many other casts already in play, we need to do this slowly and deliberately so that, to the best of my knowledge, all callers of paint_pattern are now passing their true type, rather than one which it has already been cast to.

* Change rect_draw_some_item to take a destination parameter of type HBITMAP or HDC

From quadruple+ reading of the code, this "dest" parameter can only have either a HDC (main_win == 2) or HBITMAP (main_win == 0 or 1), so use a variant to represent that.
This change is slightly dishonest as we're still pretending in the helper functions that a HWND is being passed when in fact it's a HBITMAP.

* draw_dialog_graphic now takes typesafe destination parameter

The helpers are still lying in name: for example rect_draw_some_item_wnd in fact will throw an exception if a HWND is actually contained in the variant. This will be corrected shortly.

* Replace final reinterpret_cast of HBITMAP/HDC/HWND

With this final example changed, we're now using variant types, albeit in a primitive way to provide type safe variant behaviour.

* Simplify draw_dest ahead of overall simplification

* Move dest_draw check in place so that we can possibly simplify further

* Replace pattern or rect_draw_some_item_wnd/rect_draw_some_item_dc with rect_draw_some_item_either

The pattern goes from:

```if (win_or_gworld == 1)
rect_draw_some_item_wnd(from_gworld,from_rect,hDlg,small_monst_rect,0,0);
else rect_draw_some_item_dc(from_gworld,from_rect,hdc,small_monst_rect,0, (win_or_gworld == 1) ? 0 : 2);
```
To:
`rect_draw_some_item_either(from_gworld,from_rect,win_or_gworld,hDlg,hdc,rect,0, (win_or_gworld == 1) ? 0 : 2);`

Again, this should be a stopgap measure since the HWND is really an HBITMAP here.

* Remove all traces of rect_draw_some_item_either

We can now call our variant aware version with either the variant type itself or a supported parameter.