Effect pattern global refactor
The existing model was to have a copy of an `effect_pat_type` called `current_pat` which was overwritten according to a choice of new pattern. Since realising this pattern is not updated, I made this a reference to the particular pattern choice, which can then be re-assigned to a different pattern choice as needed.