Global management
Until we can do away with the way it currently works, we need a place to keep all the nastiness in one place - and this will differ between the game and the scenario editor, and possible the party editor.
So for now we'll create a home for these global objects per project (well, not yet the party).