Remove redundant checks
All these functions start by checking that the player character status is normal, and if it isn't they return. However they then later check that the status is normal before doing some further work. In none of these cases does any operation happen between the first check and the second that would change that status, so it's redundant (and confusing!).