Add damage_type enum to make code more readable
I actually did this as a scoped enum initially in order to track down all the types, which is why there is some temporarily commented out code.
I reverted the scoped part temporarily just so that I could compare the before after text/rdata sections to see that nothing had changed as this is much easier that checking every single line where I've replaced a value with an enum.
After this checkin I will turn it back into a scoped enum.