Make damage_type a scoped enum and provide overloads of functions
This allows us to pass the scoped enum where we can. In the next change this will be improved type safety further by removing some of the to_underlying cases, but this is done in two steps to make it easier to review.
Also now that damage_type is scoped I can rename PoisonX back to Poison and it won't conflict with the other unscoped enum with the same value name.