Refactoring 7 (#52)
* Tidy up global/static variable declartions and remove any unnecessary

Also move `which_pc_displayed` definition into the only file that actually uses it and make it static.

* Simplify party struct initialization by zero initializing it

* Correct silly mistake made when assert was first added

And *this* is an obvious example of why many always like to bracket their if/else/for/while statements!

* Reformat party.cpp adding braces to if/for statements

Compared text/rdata sections before/after these changes to confirm that they are identical (i.e. no accidental changes)

* Remove redundant checks

All these functions start by checking that the player character status is normal, and if it isn't they return. However they then later check that the status is normal before doing some further work. In none of these cases does any operation happen between the first check and the second that would change that status, so it's redundant (and confusing!).

* Try to separate the logic involved when a PC is afflicted by something into separate functions to those that report/reflect the outcome

These can then be moved to pc.hpp/.cpp

* Move player affecting logic to pc.hpp/pc.cpp

* Move pc affect functions to leverage C++17's `if` init-statements

This reduces local state scope.

* Move on_spell_menu global to become a file static in party.cpp and change type to std::array

* Add enum to represent damage type

And reformat damage_monst ahead of some changes. Compilation showed no changes to text/rdata sections

* Add damage_type enum to make code more readable

I actually did this as a scoped enum initially in order to track down all the types, which is why there is some temporarily commented out code.
I reverted the scoped part temporarily just so that I could compare the before after text/rdata sections to see that nothing had changed as this is much easier that checking every single line where I've replaced a value with an enum.
After this checkin I will turn it back into a scoped enum.

* Make damage_type a scoped enum and provide overloads of functions

This allows us to pass the scoped enum where we can. In the next change this will be improved type safety further by removing some of the to_underlying cases, but this is done in two steps to make it easier to review.
Also now that damage_type is scoped I can rename PoisonX back to Poison and it won't conflict with the other unscoped enum with the same value name.

* Provide common functions which allow both routes to pass a typesafe value

* Reorder functions to avoid the need to specify function prototype

* Change  victim from pointer to reference

* Rename parameter to "type"

* Reorder or remove locals

* Create function to return the hit chance of a particular skill level

Very low level for now, but hopefully more logic will be built on top of it